+++ packages/core/mashroom-utils/test/data/mashroom.ts
@@ -32,7 +32,7 @@ const serverConfig: MashroomServerConfig = {
             },
             ssrConfig: {
                 ssrEnable: true,
-                renderTimoutMs: 2000,
+                renderTimeoutMs: 2000,
                 cacheEnable: false,
                 cacheTTLSec: 300
             }