+++ packages/plugin-packages/mashroom-portal/test/backend/controllers/PortalAppController.test.ts
@@ -31,7 +31,7 @@ setPortalPluginConfig({
     defaultProxyConfig: {},
     ssrConfig: {
         ssrEnable: false,
-        renderTimoutMs: 2000,
+        renderTimeoutMs: 2000,
         cacheEnable: false,
         cacheTTLSec: 300,
         inlineStyles: true,