+++ packages/plugin-packages/mashroom-portal/test/backend/controllers/PortalPageEnhancementController.test.ts
@@ -25,7 +25,7 @@ setPortalPluginConfig({
     defaultProxyConfig: {},
     ssrConfig: {
         ssrEnable: false,
-        renderTimoutMs: 2000,
+        renderTimeoutMs: 2000,
         cacheEnable: false,
         cacheTTLSec: 300,
         inlineStyles: true,