+++ packages/plugin-packages/mashroom-portal/test/backend/utils/ssr-utils.test.ts
@@ -24,7 +24,7 @@ setPortalPluginConfig({
     defaultProxyConfig: {},
     ssrConfig: {
         ssrEnable: true,
-        renderTimoutMs: 2000,
+        renderTimeoutMs: 2000,
         cacheEnable: true,
         cacheTTLSec: 300,
         inlineStyles: true,